chore: ViewController파일을 FromView, FromModel로 분리
